Move images into the public folder and use paths instead of importing.
